+++ test/avro_schema_test.exs
@@ -269,11 +269,14 @@ defmodule AvroSchemaTest do
 
   test "atom values are encoded as strings for string types", %{schema_json: schema_json} do
     {:ok, encoder} = AvroSchema.make_encoder(schema_json)
+    {:ok, decoder} = AvroSchema.make_decoder(schema_json)
 
     atom_value = %{field1: :some_atom, field2: 2}
 
     {:ok, encoded} = AvroSchema.encode(atom_value, encoder)
-    assert is_list(encoded)
+    {:ok, decoded} = AvroSchema.decode(encoded, decoder)
+
+    assert decoded == %{"field1" => "some_atom", "field2" => 2}
   end
 
   test "subject", %{schema_json: schema_json} do