+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [0.3.0] - 2020-02-17
+### Changed
+- Pulls in [erlavro 2.9.0](https://github.com/klarna/erlavro/compare/e43fe8b..cd49689) which encodes atom values as strings for string types
 
 ## [0.2.0] - 2020-02-05
 ### Changed